organized and moved all test files to ExampleMazeTests.java, and moved the runnable game to ExampleMaze.java
